* c3005c5: Fix crucial bugs in OperationNaiveEvalHessian*
* bb793ec: Add -Wno-unused-parameter to Clang flags
* 9ddf8b5: Fix wrong include guards (thanks to Clang)
* 623afa0: Add missing override keywords (thanks to Clang)
* 42ef2db: HashGridIndex: Use friend struct (thanks to Clang)
* 83bb0f3: Remove extraneous parentheses (thanks to Clang)
* 3a5568f: Fix wrong argument name (thanks to Clang)
* 19b4391: Add more override keywords (thanks to Clang)
* ffc7dd2: Fix -Wsign-compare warning (Clang)
* cce3d02: Fix serious bugs with uninitialized vars (Clang)
* 18d526f: Fix -Wunneeded-internal-declaration warning (Clang)
* 3c2bd5f: Fix unused private field warnings (Clang)
* 9feb11c: Replace wrong deletes with delete[] (Clang)
* 48e34b2: Remove unnecessary std::moves (Clang)
* 40ec787: Remove redundant virtuals due to override (Clang)
* 3a69f71: Remove -Woverloaded-virtual warnings (Clang)
* 5be9432: Log stderr and output of child processes
* ab12e7b: Fix GCC -Wconversion warnings
* bace230: Fix MPI error
* 05757c1: Log output of child processes also on Windows
* b0b09ba: Try other method for writing to build.log
* 1a500d2: Fix screwed up commit ab12e7b
* 2d1627e: Revert 05757c1 as this caused deadlocks with -j
* b0e84e6: Fix MPI errors (follow-up to bace230)
* 39ce5e5: Fix missing import in 2d1627e

+++ parallel/src/sgpp/parallel/pde/application/PoissonEquationSolverMPI.hpp
@@ -52,7 +52,7 @@ class PoissonEquationSolverMPI : public sgpp::pde::EllipticPDESolver {
    */
   virtual ~PoissonEquationSolverMPI();
 
-  void constructGrid(sgpp::base::BoundingBox& myBoundingBox, int level);
+  void constructGrid(sgpp::base::BoundingBox& myBoundingBox, size_t level);
 
   void solvePDE(sgpp::base::DataVector& alpha, sgpp::base::DataVector& rhs, size_t maxCGIterations,
                 double epsilonCG, bool verbose = false);
